requirements.txt
transformers
discord.py[voice]
sentence_transformers
detoxify
youtube-search-python
youtube-dl